requirements.txt
pandas==2.2.3
openpyxl==3.1.5
openai==1.65.1
python-dotenv==1.0.1
tabulate==0.9.0
matplotlib==3.8.3 